backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
httpx==0.27.2
pyyaml==6.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.3
pytest==8.3.3
pytest-asyncio==0.24.0

# SSH and network connectivity
paramiko==3.4.0
asyncssh==2.14.2

# Background tasks and scheduling
celery==5.3.4
redis==5.0.1

# UniFi Controller integration
aiohttp==3.9.1
